Move training code to __main__ and drop redundant functions
